Latency estimator for #16
We probably won't end up using this, but here it is for posterity, just in case it proves useful.

Largely this didn't work because the platform automatically retires TCP connections according to its own algorithm - which totally alters the timings.  See http://stackoverflow.com/questions/22417228.

Will be backed out in a moment.